+++ src/stdlib/string.rs
@@ -10,6 +10,6 @@ function! {
     argc: Some(1),
     callback: |state, args| {
         let len = args[0].eval(state)?.string()?.len();
-        Ok(Atom::Int(len as i32))
+        Ok(Atom::Int(len as i64))
     },
 }
